[Object_Detection] Add Hard Negative Mining.
Note that this does NOT implement the caveat described in
[Training Region-based Object Detectors with Online Hard Example Mining]
(https://arxiv.org/pdf/1604.03540.pdf), section 4.1, where NMS is used
to perform deduplication. I believe this can be done before getting to
this layer, but it can also be done as part of this layer. For now,
I will prefer composibility. If NMS is ALWAYS used together with negative
mining, then we should consider adding an argument in this layer.