Update visualization/plot_image_gallery.py
Update visualization/plot_image_gallery.py 
The guide has an issue when calling the visualization.plot_image_gallery function with its input being a list, it has to be of type array because it will be calling for the shape of it later in the function. The error occurred in the second code cell in the Beginner section, Ligne 155 in the object_detection_keras_cv.py file. This update should solve it.